feat: add `Login` hook and set `authAtom`, `userAtom`
